Доработки в отрисовке больших фото
