+++ js/main.js
@@ -1,4 +1,3 @@
 import {renderThumbnails} from './render-thumbnails.js';
-import './render-big-pictures.js';
 
 renderThumbnails();
